backend/library/llm-universal/requirements.txt
# openai>=1.0.0
# langchain-core>=0.1.0
# langchain-openai>=0.0.1
# google-generativeai>=0.8.0
# tiktoken>=0.5.0
# google-api-core>=2.11.0

openai>=1.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.1
google-genai>=0.3.0
google-api-core>=2.15.0
tiktoken>=0.5.0